lab2/requirements.txt
requests>=2.28.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
python-dotenv>=0.19.0